requirements.txt
absl-py==2.3.0
AccessControl==7.2
Acquisition==6.1
aiosignal==1.3.2
altair==5.5.0
AMFM_decompy==1.0.12.2
amqp==5.3.1
annotated-types==0.7.0
apprise==1.9.3
asn1crypto==1.5.1
astral==3.2
astunparse==1.6.3
AuthEncoding==6.0
Automat==25.4.16
base58check==1.0.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
bech32==1.2.0
better-profanity==0.7.0
billiard==4.2.1
bitarray==3.4.3
blake256==0.1.1
bleach==6.2.0
blinker==1.9.0
boto3==1.38.43
botocore==1.38.43
BTrees==6.1
cbor==1.0.0
cbor2==5.6.5
celery==5.5.3
certifi==2025.6.15
Chameleon==4.6.0
chardet==3.0.4
ckzg==2.1.1
coinaddrvalidator==1.2.3
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.3.2
crepe==0.0.16
crispy-bootstrap4==2025.6
cssselect==1.3.0
cycler==0.12.1
Cython==3.1.2
cytoolz==1.0.1
DateTime==5.5
decorator==5.2.1
deepface==0.0.93
Django==5.2.3
django-address==0.2.8
django-allauth==65.9.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-db-connection-pool==1.2.6
django-extensions==4.1
django-jsignature==0.12
django-recaptcha==4.1.0
django-simple-history==3.10.1
django-storages==1.14.6
django-summernote==0.8.20.0
django-webauth==0.1.2
django-webpush==0.3.6
dnspython==2.7.0
DocumentTemplate==4.6
dsplib==0.10.1
email_validator==2.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
ExtensionClass==6.0
facexlib==0.3.0
ffmpeg==1.4
filelock==3.18.0
filterpy==1.4.5
fire==0.7.0
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-oauth==1.0.1
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googletrans==3.1.0a0
greenlet==3.2.3
groestlcoin_hash2==1.1.1
grpcio==1.73.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
h3==4.3.0
h5py==3.14.0
hexbytes==1.3.1
hf-xet==1.1.5
hmmlearn==0.3.3
hpack==3.0.0
hstspreload==2025.1.1
http_ece==1.2.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.33.0
humanfriendly==10.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
ijson==3.4.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==6.11.0
imutils==0.5.4
incremental==24.7.2
itsdangerous==2.2.0
jellyfish==1.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jwt==1.4.0
keras==3.10.0
kiwisolver==1.4.8
kombu==5.5.4
langdetect==1.0.9
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lmdb==1.6.2
lxml==5.4.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
mtcnn==1.0.0
multidict==6.5.1
MultiMapping==5.0
multipart==1.2.1
namex==0.1.0
narwhals==1.44.0
networkx==3.5
nudenet==3.4.2
numba==0.61.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauth2client==4.1.3
oauthlib==3.3.1
onnxruntime==1.22.0
openai==0.28.1
openai-clip==1.0.1
opt_einsum==3.4.0
optree==0.16.0
packaging==23.2
pandas==2.3.0
parsimonious==0.10.0
PasteDeploy==3.1.0
pdf417==0.8.1
pedalboard==0.9.17
penn==1.0.0
Persistence==5.1
persistent==6.1.1
pillow==11.2.1
pitch-detectors==0.7.0
platformdirs==4.3.8
pooch==1.8.2
praat-parselmouth==0.4.6
priority==1.3.0
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psycopg-binary==3.2.9
psycopg2==2.9.10
pyaml==25.5.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.9.0
pycparser==2.22
pycryptodome==3.23.0
pydeck==0.9.1
pydub==0.25.1
pyFLAC==3.0.0
Pygments==2.19.2
pyheif==0.5.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyotp==2.9.0
pyparsing==3.2.3
pyquery==2.0.1
pyreaper==0.0.11
PySocks==1.7.1
pysptk==1.0.1
python-cryptapi==1.0.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-gettext==5.0
python-magic==0.4.27
pytz==2025.2
pyunormalize==16.0.0
pyworld==0.3.5
PyYAML==6.0.2
realesrgan==0.3.0
realtime-crypto==0.0.1
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
resampy==0.4.3
RestrictedPython==8.1a1.dev0
retina-face==0.0.17
rfc3986==1.5.0
rich==13.9.4
rlp==4.1.0
roman==5.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.14.0
sentry-sdk==2.31.0
sewar==0.4.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sqlparams==6.2.0
sqlparse==0.5.3
srt==3.5.3
streamlit==1.46.0
stripe==12.2.0
sympy==1.14.0
ta-lib==0.6.4
tb-nightly==2.20.0a20250624
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorrt==10.0.1
tensorrt_cu12==10.12.0.36
tensorrt_cu12_bindings==10.12.0.36
tensorrt_cu12_libs==10.12.0.36
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2025.6.11
timezonefinder==6.5.9
tokenizers==0.21.2
toml==0.10.2
toolz==1.0.0
torbi==1.1.0
torch==2.7.1
torch-yin==0.1.3
torchaudio==2.7.1
torchcrepe==0.0.24
torchutil==0.0.14
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transaction==5.0
transformers==4.52.4
triton==3.3.1
Twisted==25.5.0
txaio==23.1.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urlextract==1.9.0
validators==0.35.0
vine==5.1.0
vosk==0.3.45
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
web3==7.12.0
webencodings==0.5.1
WebOb==1.8.9
webrtcvad==2.0.10
websockets==15.0.1
WebTest==3.0.6
Werkzeug==3.1.3
whatimage==0.0.3
wheel==0.45.1
wrapt==1.17.2
WSGIProxy2==0.5.1
yapecs==0.4.0
yapf==0.43.0
yarl==1.20.1
z3c.pt==5.1
zc.lockfile==3.0.post1
ZConfig==4.2
zExceptions==5.0
zipp==3.23.0
ZODB==6.0.1
zodbpickle==4.2
Zope==5.13
zope.annotation==5.1
zope.browser==3.1
zope.browsermenu==5.1
zope.browserpage==5.0
zope.browserresource==5.2
zope.cachedescriptors==5.1
zope.component==6.0
zope.configuration==6.0
zope.container==6.1
zope.contentprovider==6.0
zope.contenttype==5.2
zope.datetime==5.1
zope.deferredimport==5.0
zope.deprecation==5.1
zope.dottedname==6.0
zope.event==5.0
zope.exceptions==5.2
zope.filerepresentation==6.1
zope.globalrequest==2.0
zope.hookable==7.0
zope.i18n==5.2
zope.i18nmessageid==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.location==5.1
zope.pagetemplate==5.1
zope.processlifetime==3.1
zope.proxy==6.1
zope.ptresource==5.1
zope.publisher==7.3
zope.schema==7.0.1
zope.security==7.3
zope.sequencesort==5.1
zope.site==5.1
zope.size==5.1
zope.structuredtext==5.0
zope.tal==5.1
zope.tales==6.1
zope.testbrowser==7.0.1
zope.testing==5.1
zope.traversing==5.1
zope.viewlet==5.1
zxing==1.0.3
setuptools==58.2.0
twilio
autocorrect
paramiko
ccxt
numpy
urllib3
webauthn
